Replace most of `std::map/set/unordered_map/unordered_set` with more efficient collections from abseil-cpp
Taken from 'memory optimizations' branch to make it more readable and
concise.

+++ worker/src/RTC/NackGenerator.cpp
@@ -152,18 +152,18 @@ namespace RTC
 		// clear it and request a keyframe.
 		uint16_t numNewNacks = seqEnd - seqStart;
 
-		if (this->nackList.size() + numNewNacks > MaxNackPackets)
+		if (static_cast<uint16_t>(this->nackList.size()) + numNewNacks > MaxNackPackets)
 		{
 			// clang-format off
 			while (
 				RemoveNackItemsUntilKeyFrame() &&
-				this->nackList.size() + numNewNacks > MaxNackPackets
+				static_cast<uint16_t>(this->nackList.size()) + numNewNacks > MaxNackPackets
 			)
 			// clang-format on
 			{
 			}
 
-			if (this->nackList.size() + numNewNacks > MaxNackPackets)
+			if (static_cast<uint16_t>(this->nackList.size()) + numNewNacks > MaxNackPackets)
 			{
 				MS_WARN_TAG(
 				  rtx, "NACK list full, clearing it and requesting a key frame [seqEnd:%" PRIu16 "]", seqEnd);
